whisperwood_villa_android_tab/requirements.txt
Flask
requests
psycopg2-binary
bcrypt
